Move MeasureText types to the utils file that uses them
