+++ frontend/src/metabase/utils/measure-text.ts
@@ -1,9 +1,27 @@
 import _ from "underscore";
 
-import type {
-  FontStyle,
-  TextMeasurer,
-} from "metabase/visualizations/shared/types/measure-text";
+export type FontStyle = {
+  size: string | number;
+  family: string;
+  weight: string | number;
+};
+
+export interface TextSize {
+  width: number;
+  height: number;
+}
+
+export type TextWidthMeasurer = (
+  text: string,
+  style: FontStyle,
+) => TextSize["width"];
+
+export type TextHeightMeasurer = (
+  text: string,
+  style: FontStyle,
+) => TextSize["height"];
+
+export type TextMeasurer = (text: string, style: FontStyle) => TextSize;
 
 let canvas: HTMLCanvasElement | null = null;
